requirements.txt
selenium==4.33.0
openai==1.91.0
python-dotenv==1.1.1
croniter>=1.3.0
pytz>=2023.3
playwright==1.35.0
async-playwright==1.35.0
requests==2.31.0
beautifulsoup4==4.12.2
